+++ lib/flutter_show_menu.dart
@@ -2,3 +2,4 @@ export 'src/menu_position.dart';
 export 'src/overlay_menu.dart' show showOverlayMenu;
 export 'src/overlay_menu_item.dart';
 export 'src/overlay_menu_button.dart';
+export 'src/overlay_menu_style.dart';